Restructure PDU code
The PDU class is now solely responsible for communication with the PDU.
It exposes public high-level communication methods that the pdu_channel can
call.
All methods are now have a return type of bool, where any failure
returns false, and true means the operation succeeded.
It is the responsibility of the calling function to retry on failure and
wait for timeout.
The pdu_channel has been broken up into smaller helper functions.
The PDU class now stores the state of all switches in an array, which
can be publicly accessed.